backend/requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.2
pydantic==2.11.3
pydantic-settings==2.8.1
python-dotenv==1.1.0
email-validator==2.2.0

SQLAlchemy==2.0.39
alembic==1.15.2
psycopg[binary]==3.2.6

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

httpx==0.28.1
